refactor(auth): AuthContext 관련 로직 파일 분리 및 구조 개선
- AuthContext 내부에 정의된 context, provider, hook을 분리하여 응집도 향상
- createContext → AuthContext.tsx
- Provider 로직 → AuthProvider.tsx
- useContext 호출 → useAuth.ts로 이동
- 관련 타입 정의 → types/auth.ts로 분리
- main.tsx에서 AuthProvider import 경로 수정
- 모든 파일들의 import 경로 수정
