+++ src/contexts/AuthContext/AuthContext.tsx
@@ -1,0 +1,6 @@
+import { createContext } from 'react';
+import type { AuthContextType } from '@/types/auth';
+
+export const AuthContext = createContext<AuthContextType | undefined>(
+  undefined
+);
