+++ src/hooks/useAuth.ts
@@ -1,0 +1,8 @@
+import { useContext } from 'react';
+import { AuthContext } from '@/contexts/AuthContext/AuthContext';
+
+export const useAuth = () => {
+  const context = useContext(AuthContext);
+  if (!context) throw new Error('useAuth는 AuthProvider 내부에서 사용해야 함');
+  return context;
+};
