+++ src/main.tsx
@@ -5,7 +5,7 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import { queryClient } from './queryClient';
 import App from '@/App';
 import { theme } from '@/styles/theme';
-import { AuthProvider } from '@/contexts/AuthContext';
+import { AuthProvider } from '@/contexts/AuthContext/AuthProvider';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
