+++ src/pages/Mypage.tsx
@@ -1,4 +1,4 @@
-import { useAuth } from '@/contexts/AuthContext';
+import { useAuth } from '@/hooks/useAuth';
 import { useNavigate } from 'react-router-dom';
 import styled from '@emotion/styled';
 import Navigation from '@/components/Navigation';
